added all necessary mappings for router

+++ src/main/java/com/bindschaedel/nachhilfe/repository/EntryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.bindschaedel.nachhilfe.DAO.Entry;
 
-public interface EntryRepository extends CrudRepository<Entry, Long>{
+public interface EntryRepository extends CrudRepository<Entry, Integer>{
 
 }
